requirements.txt
Flask
boto3
Pillow